move table process func
