+++ app/domain/adminapi/table_transfer.py
@@ -1,6 +1,7 @@
 from app.data import model, repositories
-from app.domain.adminapi import cross_identification
+from app.domain import converters
 from app.lib.storage import enums
+from app.lib.web import errors
 from app.presentation import adminapi
 
 
@@ -21,12 +22,46 @@ class TableTransferManager:
         return adminapi.TableStatusStatsResponse(processing=self.layer0_repo.get_object_statuses(r.table_id))
 
     def table_process(self, r: adminapi.TableProcessRequest) -> adminapi.TableProcessResponse:
-        return cross_identification.table_process(
-            self.layer0_repo,
-            self.layer2_repo,
-            cross_identification.cross_identification,
-            r,
-        )
+        meta = self.layer0_repo.fetch_metadata(r.table_id)
+
+        convs: list[converters.QuantityConverter] = [
+            converters.NameConverter(),
+            converters.ICRSConverter(),
+        ]
+
+        for conv in convs:
+            try:
+                conv.parse_columns(meta.column_descriptions)
+            except converters.ConverterError as e:
+                raise errors.LogicalError(f"Unable to process table: {str(e)}") from e
+
+        offset = 0
+
+        while True:
+            data = self.layer0_repo.fetch_raw_data(
+                r.table_id, order_column=repositories.INTERNAL_ID_COLUMN_NAME, limit=r.batch_size, offset=offset
+            )
+            offset += min(r.batch_size, len(data.data))
+
+            if len(data.data) == 0:
+                break
+
+            for obj_data in data.data.to_dict(orient="records"):
+                objects: list[model.CatalogObject] = []
+                for conv in convs:
+                    obj = conv.apply(obj_data)
+                    objects.append(obj)
+
+                # TODO: cross-identification
+
+                self.layer0_repo.upsert_object(
+                    r.table_id,
+                    get_processing_info(obj_data[repositories.INTERNAL_ID_COLUMN_NAME], objects),
+                )
+
+        # TODO: remove col_name from entities?
+
+        return adminapi.TableProcessResponse()
 
     def set_table_status(self, r: adminapi.SetTableStatusRequest) -> adminapi.SetTableStatusResponse:
         # TODO:
@@ -63,6 +98,15 @@ class TableTransferManager:
                 break
 
         return adminapi.SetTableStatusResponse()
+
+
+def get_processing_info(object_id: str, data: list[model.CatalogObject]) -> model.Layer0CatalogObject:
+    """
+    :param object_id: Internal ID of the object
+    :param res: Object that stores data about the cross identification processing
+    :param obj: Processed and homogenous information about the object
+    """
+    return model.Layer0CatalogObject(object_id, enums.ObjectProcessingStatus.NEW, {}, data)
 
 
 def apply_override(
